Support for Mac OS-X added
git-svn-id: svn://svn.code.sf.net/p/opengraphtheory/code/Trunk@81 7d856203-d9e4-4e18-aee0-bf315c1b057e

+++ opengt.so/Sources/transform/scale.cpp
@@ -34,7 +34,7 @@ namespace OpenGraphtheory
             for(Graph::VertexIterator v1 = G.BeginVertices(); v1 != G.EndVertices(); v1++)
             {
                 coordinates = v1.GetCoordinates();
-                for(unsigned int j = 0; j < min(dimensions, coordinates.size()); ++j)
+                for(unsigned int j = 0; j < min(dimensions, (unsigned int)(coordinates.size())); ++j)
                 {
                     if(coordinates[j] < dimension_min[j])
                         dimension_min[j] = coordinates[j];
